+++ src/events.rs
@@ -33,3 +33,7 @@ pub fn installed(id: &str, version: &str, config: &Config) {
         println!("{{\"crate\":\"{id}\",\"version\":\"{version}\",\"event_version\":\"{EVENT_VERSION}\",\"event\":\"installed\",\"data\":\"{id}@{version}\"}}");
     }
 }
+
+pub fn get_latest(id: &str, version: &str) {
+    println!("{{\"crate\":\"{id}\",\"version\":\"{version}\",\"event_version\":\"{EVENT_VERSION}\",\"event\":\"latest_version\",\"data\":\"{version}\"}}");
+}