+++ CHANGELOG.md
@@ -37,6 +37,7 @@
 - Add --index-key which will select an index based on its entry in the config file. (#78)
 - Add --config to allow selection of a config file anywhere. (#80)
 - Use panics instead of using std::process::exit. (#47)
+- Add way to get latest versions of crates. (#84)
 
 ## [0.5.3](https://github.com/cargo-prebuilt/cargo-prebuilt/releases/tag/v0.5.3)